Refactor AsyncBeacon and RawStreamBeacon classes to abstract shared functionality
