+++ packages/state_beacon_core/test/src/beacons/future_test.dart
@@ -451,7 +451,6 @@ void main() {
     await expectLater(
       stats.stream,
       emitsInOrder([
-        isA<AsyncData<String>>(),
         isA<AsyncLoading<String>>(),
         isA<AsyncError<String>>(),
       ]),
